+++ iwopy/interfaces/pymoo/problem.py
@@ -197,4 +197,4 @@ class Pymoo_problem(Problem):
             if verbosity:
                 print()
 
-        return OptResults(suc, xi, xf, objs, cons, res)
+        return OptResults(self.problem, suc, xi, xf, objs, cons, res)